RMA Request WIP: Add simple redirect to RMA from Product Flow
This functionality will allow a use to quickly add a new asset if it doesn't exist when creating an RMA.
The user will click a "New" button that will redirect to productflow.receiving with the new_rma=1 paramater passed on the url.
This then sets a session variable that gets checked when the asset is saved and will redirect back to the RMA create page with the new asset pre-selected.

TODO:
- Add "New" button to rma create form to redirect to productflow.receiving.

Signed-off-by: Ryan Huntington <rhuntington719@gmail.com>